Move sfdxProjectJsonParser
