+++ packages/salesforcedx-vscode-core/src/sfdxProject/sfdxProjectJsonParser.ts
